Fix dict serde for complex assessment objects
