+++ mlflow/entities/trace_info.py
@@ -115,7 +115,9 @@ class TraceInfo(_MlflowObject):
     def from_proto(cls, proto) -> "TraceInfo":
         return cls(
             trace_id=proto.trace_id,
-            client_request_id=proto.client_request_id,
+            client_request_id=(
+                proto.client_request_id if proto.HasField("client_request_id") else None
+            ),
             trace_location=TraceLocation.from_proto(proto.trace_location),
             request_preview=proto.request_preview,
             response_preview=proto.response_preview,
